Added (Still BAD integer support for AVX-1 platforms) Added JSON support Added configurable dispatcher, scheduler, execution engine Fixed a few AVX2 computation bugs.

+++ paranoise/paranoise/parallel/float4.h
@@ -19,9 +19,9 @@ namespace zzsystems { namespace simdal {
 
 		float4() = default;
 		
-		float4(const int rhs)		: val(_mm_set1_ps(rhs)) {}
+		float4(const int rhs)		: val(_mm_set1_ps(static_cast<float>(rhs))) {}
 		float4(const float rhs)		: val(_mm_set1_ps(rhs)) {}
-		float4(const double rhs)	: val(_mm_set1_ps(rhs)) {}
+		float4(const double rhs)	: val(_mm_set1_ps(static_cast<float>(rhs))) {}
 
 		float4(const float* rhs)	: val(_mm_load_ps(rhs)) {}
